# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Serein API
  version: "2.1.0"
servers:
  - url: http://127.0.0.1:12510
paths:
  # Tracker
  /api/tracker/list/all:
    get:
      tags: [Tracker]
      summary: 列出全部 Tracker
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackerInfo"
  /api/tracker/list/{tracker_id}:
    get:
      tags: [Tracker]
      summary: 列出某个 Tracker 的条目
      parameters:
        - name: tracker_id
          in: path
          required: true
          description: 不含 .toml 后缀的 Tracker 文件名
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackerDetail"
  /api/tracker/new:
    post:
      tags: [Tracker]
      summary: 新建 Tracker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
      responses:
        "201": { description: 已创建 }
        "409": { description: id 已存在 }
  /api/tracker/add:
    post:
      tags: [Tracker]
      summary: 追加条目到默认 Tracker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_id]
              properties:
                app_id: { type: string }
                platforms: { type: array, items: { type: string } }
      responses:
        "200": { description: OK }
  /api/tracker/apps:
    get:
      tags: [Tracker]
      summary: 列出全部已追踪的 app_id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { type: string }

  # Check
  /api/check:
    post:
      tags: [Check]
      summary: 发起检查
      description: |
        检查 Tracker，异步执行。scope 为 all 时检查全部，为 trackers 时检查
        tracker_ids 指定的 Tracker，为 ids 时按 Tracker 分组检查给定 id。
        进度见 /api/progress/{task_id}。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scope]
              properties:
                scope:
                  type: string
                  enum: [all, trackers, ids]
                tracker_ids:
                  type: array
                  items: { type: string }
                  description: scope 为 trackers 时必填
                ids:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                  description: scope 为 ids 时必填，key 为 Tracker 名。
              example: { scope: "all" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/check/confirm:
    post:
      tags: [Check]
      summary: 确认版本号
      description: |
        一次确认一个应用。app_id 为应用 ID，其余为形如 "平台名:版本号" 的键值对。
        例：{"app_id": "pub.sakura.Zen", "macos": "1.2.3"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_id]
              additionalProperties: { type: string }
              properties:
                app_id: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  app_id: { type: string }
                  status: { type: string }
                  platforms:
                    type: object
                    description: 确认后的版本号，含服务端写入的 _confirmed_at
                    additionalProperties: { type: string }
        "400": { description: 缺少 app_id }
        "500": { description: 写入失败 }
  /api/check/result/{tracker_id}:
    get:
      tags: [Check]
      summary: 读取检查结果缓存
      description: 结果只存在于内存中，进程重启后为空。
      parameters:
        - name: tracker_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/CheckResponse"

  # Progress (SSE)
  /api/progress/{task_id}:
    get:
      tags: [Progress]
      summary: 任务进度流
      responses:
        "200":
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  每行 data: <json>。每个任务以 step 为 done 的事件结束。

                  同步过程中为 start、list、source、skip、file、write、error、done。
                  检查过程中为 start、app、done。

                  done 与 total 跨源累计，只增不减。
                  start 只发一次。
                  files 为本轮实际下载的文件数，跳过的不计入。
        "404": { description: 任务不存在或已过期 }
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
  /api/progress/cancel/{task_id}:
    post:
      tags: [Progress]
      summary: 终止任务
      description: 任务自行收尾，已完成的部分将保留。
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 已取消 }
        "404": { description: 任务不存在 }

  # Events (SSE)
  /api/events:
    get:
      tags: [Events]
      summary: 全局事件流
      description: |
        订阅时会补发历史事件。error 级别的提示常驻，warn 级别的 8 秒后消失。
      responses:
        "200":
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  data: {"level":"error","context":"[check]","message":"...","sticky":true,"ts":1720000000}

  # Config
  /api/config:
    get:
      tags: [Config]
      summary: 读取当前配置
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
                    description: 面向展示的键值对，含监听地址、下载器、并发数、平台、规则源。
                  profile:
                    type: object
                    description: 动态配置，读取失败时为空对象。
                    properties:
                      known_extensions:
                        type: array
                        items: { type: string }
                      version_prefixes:
                        type: array
                        items: { type: string }
                      version_suffixes:
                        type: array
                        items: { type: string }
                  first_run:
                    type: boolean
                  rule_source_dev:
                    type: string
                    description: 本地规则目录，为空表示未启用。

  # Sync
  /api/sync:
    post:
      tags: [Sync]
      summary: 拉取规则或动态配置
      description: |
        type 为 rules 时返回 task_id，进度见 /api/progress/{task_id}。
        type 为 profile 时直接返回结果。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [rules, profile]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TaskResponse"
                  - $ref: "#/components/schemas/ProfileSyncResponse"

  # Rules
  /api/rules:
    get:
      tags: [Rules]
      summary: 列出规则源或筛选规则
      description: |
        无参数时返回规则源列表，带 q 或 source 时返回规则列表。
        同时传 q 与 source 时，以 q 为准。q 不限制返回条数。
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: 按名称筛选
        - name: source
          in: query
          schema: { type: string }
          description: 按 source_id 筛选
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/SourceItem"
                  - type: array
                    items:
                      $ref: "#/components/schemas/RuleItem"
  /api/rules/check:
    post:
      tags: [Rules]
      summary: 校验全部规则
      description: 每次调用都会重新解析。dev 为 true 时改为校验本地规则目录。
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dev:
                  type: boolean
                  description: 为 true 时校验 rule_source_dev 指向的目录
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  issues:
                    type: array
                    items: { $ref: "#/components/schemas/RuleIssue" }
        "400": { description: dev 为 true 但未配置 rule_source_dev }

  # Search
  /api/search:
    get:
      tags: [Search]
      summary: 搜索规则与 Tracker 条目
      description: apps 合计最多 30 条，rules 不受此限制。
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: 关键词，对名称、app_id、描述做不区分大小写的子串匹配。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  apps:
                    type: array
                    items: { $ref: "#/components/schemas/SearchAppHit" }
                  rules:
                    type: array
                    items: { $ref: "#/components/schemas/RuleItem" }

  # Download
  /api/file:
    get:
      tags: [Download]
      summary: 带签名的下载代理
      description: |
        转发上游并按 name 命名文件。四个查询参数缺一不可，sig 为 HMAC-SHA256，
        密钥取自进程随机值，按 url、name、exp 三行拼接后计算。

        签名有效期 24 小时，重启后全部失效。
        目标地址会经过私网拦截。
        上游 4xx 原样返回，5xx 转为 502，同时推送一条 [下载] 事件。
      parameters:
        - name: url
          in: query
          required: true
          description: 上游地址
          schema: { type: string }
        - name: name
          in: query
          description: 落盘文件名，为空时取上游 Content-Disposition 或 URL 末段。
          schema: { type: string }
        - name: exp
          in: query
          required: true
          description: 过期时间，Unix 秒。
          schema: { type: string }
        - name: sig
          in: query
          required: true
          description: 十六进制签名字符串
          schema: { type: string }
      responses:
        "200":
          description: 上游文件内容
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "206": { description: 上游范围响应 }
        "400": { description: 缺少 url，或非 http(s) 地址 }
        "403": { description: 签名无效或已过期，或目标地址被私网拦截 }
        "502": { description: 上游请求失败或返回 5xx }
  /api/call/downloader:
    post:
      tags: [Download]
      summary: 发送到下载器
      description: 按 downloader 配置分发。下载器未运行时回退到浏览器。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, description: http(s) 绝对地址 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  message: { type: string }
        "400": { description: 缺少 url，或非 http(s) 地址 }
        "403": { description: 仅本机可用 }
  /api/call/browser:
    post:
      tags: [Download]
      summary: 在浏览器中打开地址
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, description: http(s) 绝对地址 }
      responses:
        "200": { description: OK }
        "400": { description: 缺少 url，或非 http(s) 地址 }
        "403": { description: 仅本机可用 }

components:
  schemas:
    TaskResponse:
      type: object
      properties:
        task_id: { type: string }
        total: { type: string }
    ProfileSyncResponse:
      type: object
      properties:
        updated: { type: boolean }
        known_extensions:
          type: array
          items: { type: string }
    TrackerInfo:
      type: object
      properties:
        id: { type: string }
        display_name: { type: string }
        order: { type: integer }
        type: { type: string, description: "app | msvsix | openvsx" }
        count: { type: integer, description: 条目数 }
        updated: { type: integer, description: 有更新的条目数 }
        checked: { type: boolean, description: 是否检查过 }
    TrackerDetail:
      type: object
      properties:
        app_id: { type: string }
        name: { type: string }
        description: { type: string }
        official_website: { type: string }
        source_id: { type: string }
        source_name: { type: string }
        current_version:
          type: object
          additionalProperties: { type: string }
        confirmed_at: { type: string, description: 上次确认的 UTC 时间戳 }
        rule_missing: { type: boolean }
        status:
          $ref: "#/components/schemas/RuleStatus"
    CheckResponse:
      type: object
      properties:
        app_id: { type: string }
        name: { type: string }
        official_website: { type: string }
        platforms:
          type: object
          additionalProperties:
            type: object
            properties:
              current_version: { type: string }
              latest_version: { type: string }
              url:
                oneOf:
                  - type: string
                  - type: array
                    items: { type: string }
              error: { type: string }
              download_method: { type: string, enum: [browser, downloader] }
              download_via_proxy: { type: boolean }
              download_name: { type: string }
              proxy_url:
                oneOf:
                  - type: string
                  - type: array
                    items: { type: string }
    SearchAppHit:
      type: object
      properties:
        app_id: { type: string }
        name: { type: string }
        description: { type: string }
        tracker_id: { type: string }
        tracker_name: { type: string }
        current_version:
          type: object
          additionalProperties: { type: string }
          description: 已确认的版本号
    RuleItem:
      type: object
      properties:
        app_id: { type: string }
        name: { type: string }
        description: { type: string }
        official_website: { type: string }
        platforms:
          type: array
          items: { type: string }
        source_id: { type: string }
        source_name: { type: string }
        status:
          $ref: "#/components/schemas/RuleStatus"
    RuleStatus:
      type: object
      description: 规则状态
      properties:
        message: { type: string }
        level: { type: string, description: "warn | error | removed" }
    RuleIssue:
      type: object
      properties:
        level:
          type: string
          enum: [warn, error]
        message: { type: string }
    SourceItem:
      type: object
      properties:
        source_id: { type: string }
        name: { type: string }